@@ -1,6 +1,9 @@
 {
   "dependencies": {
-    "lodash-es": "^4.17.4",
-    "react": "^15.5.4"
+    "lodash-es": "^4.17.14",
+    "react": "^17.0.1"
+  },
+  "devDependencies": {
+    "babel-loader": "^8.2.2"
   }
 }
