@@ -1,80 +1,84 @@
 {
   "name": "thread-loader",
-  "version": "1.1.2",
-  "description": "Runs the following loaders in a worker pool.",
-  "author": "Tobias Koppers @sokra",
+  "version": "3.0.1",
+  "description": "Runs the following loaders in a worker pool",
   "license": "MIT",
+  "repository": "webpack-contrib/thread-loader",
+  "author": "Tobias Koppers @sokra",
+  "homepage": "https://github.com/webpack-contrib/thread-loader",
+  "bugs": "https://github.com/webpack-contrib/thread-loader/issues",
+  "funding": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/webpack"
+  },
   "main": "dist/cjs.js",
+  "engines": {
+    "node": ">= 10.13.0"
+  },
+  "scripts": {
+    "start": "npm run build -- -w",
+    "clean": "del-cli dist",
+    "prebuild": "npm run clean",
+    "build": "cross-env NODE_ENV=production babel src -d dist --copy-files",
+    "commitlint": "commitlint --from=master",
+    "security": "npm audit",
+    "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css,ts}\" --list-different",
+    "lint:js": "eslint --cache .",
+    "lint": "npm-run-all -l -p \"lint:**\"",
+    "test:only": "cross-env NODE_ENV=test jest --forceExit",
+    "test:watch": "npm run test:only -- --watch",
+    "test:coverage": "npm run test:only -- --collectCoverageFrom=\"src/**/*.js\" --coverage",
+    "pretest": "npm run lint",
+    "test": "npm run test:coverage",
+    "prepare": "npm run build",
+    "release": "standard-version",
+    "defaults": "webpack-defaults"
+  },
   "files": [
     "dist"
   ],
-  "scripts": {
-    "test": "jest",
-    "webpack-defaults": "webpack-defaults",
-    "start": "npm run build -- -w",
-    "build": "cross-env NODE_ENV=production babel src -d dist --ignore 'src/**/*.test.js'",
-    "clean:dist": "del-cli dist",
-    "lint": "eslint --cache src test",
-    "lint-staged": "lint-staged",
-    "prebuild": "npm run clean",
-    "prepublish": "npm run build",
-    "release": "standard-version",
-    "security": "nsp check",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --collectCoverageFrom='src/**/*.js' --coverage",
-    "travis:coverage": "npm run test:coverage -- --runInBand",
-    "travis:lint": "npm run lint && npm run security",
-    "travis:test": "npm run test -- --runInBand",
-    "appveyor:test": "npm run test",
-    "clean": "del-cli dist"
+  "peerDependencies": {
+    "webpack": "^4.27.0 || ^5.0.0"
   },
   "dependencies": {
-    "async": "^2.3.0",
-    "loader-runner": "^2.3.0",
-    "loader-utils": "^1.1.0"
+    "loader-runner": "^4.1.0",
+    "loader-utils": "^2.0.0",
+    "neo-async": "^2.6.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-jest": "^19.0.0",
-    "babel-loader": "^6.4.1",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.4.0",
-    "cross-env": "^4.0.0",
-    "css-loader": "^0.28.4",
-    "del-cli": "^0.2.1",
-    "eslint": "^3.19.0",
-    "eslint-config-webpack": "^1.2.1",
-    "eslint-plugin-import": "^2.2.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "jest": "^19.0.2",
-    "lint-staged": "^3.4.0",
-    "lodash": "^4.17.4",
-    "node-sass": "^4.5.3",
-    "nodemon": "^1.11.0",
-    "nsp": "^2.6.3",
-    "pre-commit": "^1.2.2",
-    "sass-loader": "^6.0.6",
-    "standard-version": "^4.0.0",
-    "webpack": "^3.6.0",
-    "webpack-defaults": "^1.6.0"
+    "@babel/cli": "^7.12.1",
+    "@babel/core": "^7.12.3",
+    "@babel/preset-env": "^7.12.1",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
+    "@webpack-contrib/defaults": "^6.3.0",
+    "@webpack-contrib/eslint-config-webpack": "^3.0.0",
+    "babel-jest": "^26.6.1",
+    "babel-loader": "^8.1.0",
+    "cross-env": "^7.0.2",
+    "css-loader": "^5.0.0",
+    "del": "^5.1.0",
+    "del-cli": "^3.0.1",
+    "eslint": "^7.12.1",
+    "eslint-config-prettier": "^6.14.0",
+    "eslint-plugin-import": "^2.22.1",
+    "husky": "^5.0.9",
+    "jest": "^26.6.1",
+    "lint-staged": "^10.5.0",
+    "lodash": "^4.17.20",
+    "memfs": "^3.2.0",
+    "mini-css-extract-plugin": "^1.2.1",
+    "nodemon": "^2.0.6",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.1.2",
+    "sass": "^1.32.8",
+    "sass-loader": "^10.0.4",
+    "standard-version": "^9.0.0",
+    "webpack": "^5.23.0",
+    "webpack-cli": "^4.1.0",
+    "webpack-defaults": "^4.0.0"
   },
-  "engines": {
-    "node": ">= 4.8 < 5.0.0 || >= 5.10"
-  },
-  "peerDependencies": {
-    "webpack": "^2.0.0 || ^3.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:webpack-contrib/thread-loader.git"
-  },
-  "pre-commit": "lint-staged",
-  "lint-staged": {
-    "*.js": [
-      "eslint --fix",
-      "git add"
-    ]
-  }
+  "keywords": [
+    "webpack"
+  ]
 }
