@@ -1,12 +1,10 @@
 const path = require('path');
 
-// eslint-disable-next-line import/no-extraneous-dependencies
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin'); // eslint-disable-line import/no-extraneous-dependencies
 
-// eslint-disable-line import/no-extraneous-dependencies
-const threadLoader = require('../dist');
+const threadLoader = require('../../src'); // eslint-disable-line import/no-extraneous-dependencies
 
-module.exports = (env = {}) => {
+module.exports = (env) => {
   const workerPool = {
     workers: +env.threads,
     poolTimeout: env.watch ? Infinity : 2000,
@@ -16,16 +14,17 @@ module.exports = (env = {}) => {
     workerParallelJobs: 2,
     poolTimeout: env.watch ? Infinity : 2000,
   };
-
+  const sassLoaderOptions = {
+    sourceMap: true,
+  };
   if (+env.threads > 0) {
     threadLoader.warmup(workerPool, ['babel-loader', 'babel-preset-env']);
     threadLoader.warmup(workerPoolSass, ['sass-loader', 'css-loader']);
   }
-
   return {
-    mode: 'development',
+    mode: 'none',
     context: __dirname,
-    entry: ['react', 'lodash-es', './index.js'],
+    entry: ['./index.js'],
     output: {
       path: path.resolve('dist'),
       filename: 'bundle.js',
@@ -36,7 +35,7 @@ module.exports = (env = {}) => {
           test: /\.js$/,
           use: [
             env.threads !== 0 && {
-              loader: path.resolve(__dirname, '../dist/index.js'),
+              loader: path.resolve(__dirname, '../../dist/index.js'),
               options: workerPool,
             },
             'babel-loader',
@@ -47,11 +46,11 @@ module.exports = (env = {}) => {
           use: [
             MiniCssExtractPlugin.loader,
             env.threads !== 0 && {
-              loader: path.resolve(__dirname, '../dist/index.js'),
+              loader: path.resolve(__dirname, '../../dist/index.js'),
               options: workerPoolSass,
             },
             'css-loader',
-            'sass-loader',
+            { loader: 'sass-loader', options: sassLoaderOptions },
           ].filter(Boolean),
         },
       ],
